Started splitting autocomplete items in to custom hooks
